Change edge representation.
Previously, an edge was represented as (target node, position of
target node in out edge list of source). Now it is
(source node, position of source node in in edge list of target).

Why the change? Well, we plan on representing edge properties
just as we do the self._out_nodes (i.e. a list of lists). With
the new edge encoding, an edge can directly be used to index
into a property list whose layout matches that of self._out_nodes
but which contains the property instead of the node index.

We insist on following self._out_nodes and not self._in_nodes
because we mostly enumerate things in order of _out_nodes and
this scheme is better from a CPU cache point of view.